@@ -3,9 +3,9 @@ require 'spec_helper'
 describe "HealthController" do
 
   it "should check database and return ok" do
-  	Ping.stub(:create).and_return(Ping.new)
+  	expect(Ping).to receive(:create).and_return(Ping.new)
   	get '/health'
-    last_response.should be_ok
+    expect(last_response).to be_ok
   end
 
   it 'should version should include BUILD_NUMBER' do
