@@ -3,18 +3,18 @@ require 'spec_helper'
 describe "UsersController" do
 
 	describe 'get :register' do
-		
+
 	  it "should response ok" do
-	  	JobVacancy::App.any_instance.should_receive(:render).with('users/new')
+	  	expect_any_instance_of(JobVacancy::App).to receive(:render).with('users/new')
 	  	get '/register'
-	    last_response.should be_ok
+	    expect(last_response).to be_ok
 	  end
 
 	  it 'should render users/new' do
-	  	JobVacancy::App.any_instance.should_receive(:render).with('users/new')
+	  	expect_any_instance_of(JobVacancy::App).to receive(:render).with('users/new')
 	  	get '/register'
 	  end
 
 	 end
-	 
+
 end
