@@ -6,8 +6,8 @@ describe JobApplication do
 
 		subject { @job_offer = JobApplication.new }
 
-		it { should respond_to( :applicant_email ) }
-		it { should respond_to( :job_offer) }
+		it { is_expected.to respond_to( :applicant_email ) }
+		it { is_expected.to respond_to( :job_offer) }
 
 	end
 
@@ -17,13 +17,13 @@ describe JobApplication do
 	  it 'should set applicant_email' do
 	  	email = 'applicant@test.com'
 	  	ja = JobApplication.create_for(email, JobOffer.new)
-	  	ja.applicant_email.should eq email
+	  	expect(ja.applicant_email).to eq(email)
 	  end
 
 	  it 'should set job_offer' do
 	  	offer = JobOffer.new
 	  	ja = JobApplication.create_for('applicant@test.com', offer)
-	  	ja.job_offer.should eq offer
+	  	expect(ja.job_offer).to eq(offer)
 	  end
 
 	end
@@ -35,7 +35,7 @@ describe JobApplication do
 
 	  it 'should deliver contact info notification' do
 	  	ja = JobApplication.create_for('applicant@test.com', JobOffer.new)
-	  	JobVacancy::App.should_receive(:deliver).with(:notification, :contact_info_email, ja)
+	  	expect(JobVacancy::App).to receive(:deliver).with(:notification, :contact_info_email, ja)
 	  	ja.process
 	  end
 
