@@ -1,14 +1,14 @@
 JobVacancy::App.controllers :job_offers do
-  
+
   get :my do
     @offers = JobOffer.find_by_owner(current_user)
     render 'job_offers/my_offers'
-  end    
+  end
 
   get :index do
     @offers = JobOffer.all_active
     render 'job_offers/list'
-  end  
+  end
 
   get :new do
     @job_offer = JobOffer.new
@@ -21,26 +21,26 @@ JobVacancy::App.controllers :job_offers do
   end
 
   get :edit, :with =>:offer_id  do
-    @job_offer = JobOffer.get(params[:offer_id])
+    @job_offer = JobOffer.with_pk(params[:offer_id])
     # ToDo: validate the current user is the owner of the offer
     render 'job_offers/edit'
   end
 
   get :apply, :with =>:offer_id  do
-    @job_offer = JobOffer.get(params[:offer_id])
+    @job_offer = JobOffer.with_pk(params[:offer_id])
     @job_application = JobApplication.new
     # ToDo: validate the current user is the owner of the offer
     render 'job_offers/apply'
   end
 
   post :search do
-    @offers = JobOffer.all(:title.like => "%#{params[:q]}%")
+    @offers = JobOffer.where(Sequel.like(:title, "%#{params[:q]}%"))
     render 'job_offers/list'
   end
 
 
   post :apply, :with => :offer_id do
-    @job_offer = JobOffer.get(params[:offer_id])    
+    @job_offer = JobOffer.with_pk(params[:offer_id])
     applicant_email = params[:job_application][:applicant_email]
     @job_application = JobApplication.create_for(applicant_email, @job_offer)
     @job_application.process
@@ -60,11 +60,11 @@ JobVacancy::App.controllers :job_offers do
     else
       flash.now[:error] = 'Title is mandatory'
       render 'job_offers/new'
-    end  
+    end
   end
 
   post :update, :with => :offer_id do
-    @job_offer = JobOffer.get(params[:offer_id])
+    @job_offer = JobOffer.with_pk(params[:offer_id])
     @job_offer.update(params[:job_offer])
     if @job_offer.save
       flash[:success] = 'Offer updated'
@@ -72,11 +72,11 @@ JobVacancy::App.controllers :job_offers do
     else
       flash.now[:error] = 'Title is mandatory'
       render 'job_offers/edit'
-    end  
+    end
   end
 
   put :activate, :with => :offer_id do
-    @job_offer = JobOffer.get(params[:offer_id])
+    @job_offer = JobOffer.with_pk(params[:offer_id])
     @job_offer.activate
     if @job_offer.save
       flash[:success] = 'Offer activated'
@@ -84,11 +84,11 @@ JobVacancy::App.controllers :job_offers do
     else
       flash.now[:error] = 'Operation failed'
       redirect '/job_offers/my'
-    end  
+    end
   end
 
   delete :destroy do
-    @job_offer = JobOffer.get(params[:offer_id])
+    @job_offer = JobOffer.with_pk(params[:offer_id])
     if @job_offer.destroy
       flash[:success] = 'Offer deleted'
     else
