@@ -1,15 +1,15 @@
 JobVacancy::App.helpers do
-	
+
 	def current_user=(user)
-		@current_user = user 
+		@current_user = user
 	end
 
 	def current_user
-		@current_user ||= User.find_by_id(session[:current_user])
+		@current_user ||= User.with_pk(session[:current_user])
 	end
-	
+
 	def sign_in(user)
-		session[:current_user] = user.id 
+		session[:current_user] = user.id
 		self.current_user = user
 	end
 
