@@ -1,14 +1,14 @@
-migration 2, :create_job_offers do
+Sequel.migration do
   up do
-    create_table :job_offers do
-      column :id, Integer, :serial => true
-      column :title, DataMapper::Property::String, :length => 255
-      column :location, DataMapper::Property::String, :length => 255
-      column :description, DataMapper::Property::String, :length => 255
+    create_table(:job_offers) do
+      primary_key :id
+      String :title
+      String :location
+      String :description
     end
   end
 
   down do
-    drop_table :job_offers
+    drop_table(:job_offers)
   end
 end
