@@ -1,15 +1,11 @@
-migration 4, :add_timespam_to_job_offers do
+Sequel.migration do
   up do
-  	modify_table :job_offers do
-      add_column :created_on, Date
-      add_column :updated_on, Date
-    end
+    add_column :job_offers, :created_on, Date
+    add_column :job_offers, :updated_on, Date
   end
 
   down do
-  	modify_table :job_offers do
-      drop_column :created_on
-      drop_column :updated_on
-    end
+    drop_column :job_offers, :created_on
+    drop_column :job_offers, :updated_on
   end
 end
